requirements.txt
streamlit
scikit-learn
tensorflow
pandas
numpy
matplotlib
pillow
